server/requirements.txt
dnspython
pymongo
motor
aiohttp
pyyaml